pyproject.toml
[project]
name = "lunar-car-api"
version = "0.1.0"
description = "Phase 1 FastAPI backend for Lunar Car"
requires-python = ">=3.11"
dependencies = [
  "aiosqlite>=0.20.0",
  "alembic>=1.13.1",
  "asyncpg>=0.29.0",
  "fastapi>=0.115.0",
  "greenlet>=3.0.3",
  "httpx>=0.27.0",
  "pydantic-settings>=2.2.1",
  "PyJWT>=2.8.0",
  "SQLAlchemy[asyncio]>=2.0.29",
  "uvicorn[standard]>=0.30.0",
]

[project.optional-dependencies]
test = [
  "pytest>=8.1.1",
  "pytest-asyncio>=0.23.6",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"
